Add access control to memberships + fix tests

+++ typescript/db/src/resolvers/membership.ts
@@ -1,3 +1,4 @@
+import { Context } from "@labelflow/common-resolvers";
 import {
   MutationCreateMembershipArgs,
   MutationDeleteMembershipArgs,
@@ -6,9 +7,18 @@ import {
   QueryMembershipsArgs,
 } from "@labelflow/graphql-types";
 import { Prisma } from "@prisma/client";
+import {
+  checkUserAccessToMembership,
+  checkUserAccessToWorkspace,
+} from "../repository/access-control";
 import { prisma } from "../repository/prisma-client";
 
-const membership = async (_: any, args: QueryMembershipArgs) => {
+const membership = async (
+  _: any,
+  args: QueryMembershipArgs,
+  { user }: Context
+) => {
+  await checkUserAccessToMembership({ where: args.where, user });
   const membershipFromDb = await prisma.membership.findUnique({
     where: { id: args.where.id },
   });
@@ -20,28 +30,51 @@ const membership = async (_: any, args: QueryMembershipArgs) => {
   return membershipFromDb;
 };
 
-const memberships = async (_: any, args: QueryMembershipsArgs) => {
+const memberships = async (
+  _: any,
+  args: QueryMembershipsArgs,
+  { user }: Context
+) => {
   return await prisma.membership.findMany({
+    where: { userId: user?.id },
     orderBy: { createdAt: Prisma.SortOrder.asc },
     skip: args?.skip ?? undefined,
     take: args.first ?? undefined,
   });
 };
 
-const createMembership = async (_: any, args: MutationCreateMembershipArgs) => {
+const createMembership = async (
+  _: any,
+  args: MutationCreateMembershipArgs,
+  { user }: Context
+) => {
+  await checkUserAccessToWorkspace({
+    user,
+    where: { slug: args.data.workspaceSlug },
+  });
   return await prisma.membership.create({
     data: { ...args.data, id: args.data?.id ?? undefined },
   });
 };
 
-const updateMembership = async (_: any, args: MutationUpdateMembershipArgs) => {
+const updateMembership = async (
+  _: any,
+  args: MutationUpdateMembershipArgs,
+  { user }: Context
+) => {
+  await checkUserAccessToMembership({ where: args.where, user });
   return await prisma.membership.update({
     where: args.where,
     data: { ...args.data, role: args.data?.role ?? undefined },
   });
 };
 
-const deleteMembership = async (_: any, args: MutationDeleteMembershipArgs) => {
+const deleteMembership = async (
+  _: any,
+  args: MutationDeleteMembershipArgs,
+  { user }: Context
+) => {
+  await checkUserAccessToMembership({ where: args.where, user });
   return await prisma.membership.delete({
     where: args.where,
   });
